requirements.txt
ansible==2.7.6
ansible-lint==3.4.23
anyconfig==0.9.7
arrow==0.13.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.6
binaryornot==0.4.4
Cerberus==1.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
click-completion==0.3.1
colorama==0.3.9
cookiecutter==1.6.0
cryptography==2.4.2
docker==3.7.0
docker-pycreds==0.4.0
fasteners==0.14.1
filelock==3.0.10
flake8==3.5.0
future==0.17.1
git-url-parse==1.1.0
idna==2.8
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.1.0
mccabe==0.6.1
molecule==2.19.0
monotonic==1.5
more-itertools==5.0.0
paramiko==2.4.2
pathspec==0.5.9
pbr==4.1.0
pexpect==4.6.0
pluggy==0.8.1
poyo==0.4.2
psutil==5.4.6
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
PyNaCl==1.3.0
pytest==4.1.1
python-dateutil==2.7.5
python-gilt==1.2.1
PyYAML==3.13
requests==2.21.0
sh==1.12.14
six==1.11.0
tabulate==0.8.2
testinfra==1.16.0
toml==0.10.0
tox==3.7.0
tree-format==0.1.2
urllib3==1.24.1
virtualenv==16.2.0
websocket-client==0.54.0
whichcraft==0.5.2
yamllint==1.11.1
